refactor: split analysis/mcp/findings_server.py into focused modules
