+++ src/quodeq/analysis/mcp/findings_server.py
@@ -2,222 +2,30 @@
 
 Protocol: JSON-RPC 2.0 over stdio, newline-delimited JSON (no Content-Length).
 No external dependencies.
+
+This module is the entry point.  Core logic lives in:
+- ``router``  -- FindingsRouter and data types
+- ``ref_scoring`` -- reference selection helpers
 """
 from __future__ import annotations
 
 import json
-import logging
 import sys
-from dataclasses import dataclass, field
 from pathlib import Path
-from typing import Protocol, TextIO, runtime_checkable
 
 from quodeq.analysis.subagents.file_queue import FileQueue
 from quodeq.analysis.mcp.args import ServerArgs, parse_args
-from quodeq.analysis.mcp.dispatch import (
-    read_message,
-    dispatch as _dispatch,
-)
-from quodeq.engine._ref_utils import ref_label as _ref_label, load_compiled_refs as _load_compiled_refs
+from quodeq.analysis.mcp.dispatch import read_message, dispatch as _dispatch
+from quodeq.engine._ref_utils import load_compiled_refs as _load_compiled_refs
 from quodeq.core.standards.refs import load_compiled_requirements as _load_compiled_requirements
 
-_FINDING_SCHEMA_VERSION = 1
-
-
-@dataclass
-class CompiledContext:
-    """Grouped compiled-standards data for finding enrichment."""
-    compiled_refs: dict[str, list[dict]] = field(default_factory=dict)
-    compiled_reqs: dict[str, dict] = field(default_factory=dict)
-    req_to_dim: dict[str, str] = field(default_factory=dict)
-    dimension: str | None = None
-    work_dir: Path | None = None
-
-
-@runtime_checkable
-class FileReader(Protocol):
-    """Abstraction for reading source file content."""
-    def __call__(self, path: Path) -> str: ...
-
-
-@runtime_checkable
-class DeduplicationStore(Protocol):
-    """Abstraction for finding deduplication state.
-
-    The default implementation uses a process-local ``set``.  For horizontal
-    scaling (multiple MCP server instances), implement this protocol with a
-    shared backend (e.g. Redis set) and pass it to ``FindingsRouter``.
-    """
-
-    def __contains__(self, key: tuple) -> bool: ...
-    def add(self, key: tuple) -> None: ...
-
-
-def _default_read_file(path: Path) -> str:
-    return path.read_text()
-
-
-class FindingsRouter:
-    """Deduplicates and enriches findings before writing to JSONL.
-
-    - Dedup by (principle, file, line, type) -- skips duplicate findings
-    - Auto-fills principle name (p) and dimension (d) from req ID
-    - Enriches req_refs from compiled standards -- LLM doesn't need to pick refs
-    - Returns feedback to the LLM so it can move on from duplicates
-    """
-
-    def __init__(
-        self,
-        output_fh: TextIO,
-        context: CompiledContext | None = None,
-        seen_store: DeduplicationStore | None = None,
-        file_reader: FileReader | None = None,
-    ):
-        ctx = context or CompiledContext()
-        self._fh = output_fh
-        self._refs = ctx.compiled_refs
-        self._reqs = ctx.compiled_reqs
-        self._dimension = ctx.dimension
-        self._req_to_dim = ctx.req_to_dim
-        self._seen: DeduplicationStore = seen_store if seen_store is not None else set()
-        self._work_dir = ctx.work_dir
-        self._read_file = file_reader or _default_read_file
-        self.counter = 0
-
-    _CONTEXT_LINES = 5
-    _SCOPE_PREVIEW_LINES = 10
-
-    def _enrich_code(self, finding: dict) -> None:
-        """Fill snippet and context by reading the source file from work_dir."""
-        if self._work_dir is None:
-            return
-        file_path = finding.get("file")
-        if not file_path:
-            return
-        try:
-            full_path = self._work_dir / file_path
-            # Path containment check: prevent traversal outside the work directory.
-            if not full_path.resolve().is_relative_to(self._work_dir.resolve()):
-                finding.setdefault("snippet", "")
-                finding.setdefault("context", "")
-                return
-            source_lines = self._read_file(full_path).splitlines()
-        except (OSError, UnicodeDecodeError):
-            finding.setdefault("snippet", "")
-            finding.setdefault("context", "")
-            return
-
-        line = finding.get("line", 0)
-        scope = finding.get("scope")
-
-        # Scope-level or line=0: store full file in snippet for expand-on-click
-        if scope or not line:
-            finding["snippet"] = "\n".join(source_lines)
-            finding["scope"] = scope or "file"
-            finding["context"] = None
-            return
-
-        # Normal line-level enrichment
-        end_line = finding.get("end_line") or line
-        if end_line < line:
-            line, end_line = end_line, line
-        # Clamp to file boundaries (1-indexed)
-        line = max(1, min(line, len(source_lines)))
-        end_line = max(line, min(end_line, len(source_lines)))
-
-        # Build snippet (the offending lines)
-        snippet_lines = source_lines[line - 1:end_line]
-        finding["snippet"] = "\n".join(snippet_lines)
-
-        # Build context with >>> markers
-        ctx_start = max(0, line - 1 - self._CONTEXT_LINES)
-        ctx_end = min(len(source_lines), end_line + self._CONTEXT_LINES)
-        context_parts = []
-        for i in range(ctx_start, ctx_end):
-            prefix = ">>> " if line - 1 <= i < end_line else ""
-            context_parts.append(f"{prefix}{source_lines[i]}")
-        finding["context"] = "\n".join(context_parts)
-
-    def _enrich(self, args: dict, finding: dict) -> None:
-        """Auto-fill principle, dimension, and refs from compiled standards."""
-        req = args.get("req")
-        if not req:
-            return
-        # Auto-fill principle name from req ID
-        if not args.get("p") and req in self._reqs:
-            finding["p"] = self._reqs[req]["principle"]
-        # Auto-fill dimension: prefer req-to-dim mapping (consolidated), fallback to single dimension
-        if not args.get("d"):
-            if req and req in self._req_to_dim:
-                finding["d"] = self._req_to_dim[req]
-            elif self._dimension:
-                finding["d"] = self._dimension
-        # Enrich with compiled standard refs
-        if req in self._refs:
-            finding["req_refs"] = _select_best_refs(
-                self._refs[req], args.get("w", ""), args.get("reason", ""),
-            )
-
-    def receive(self, args: dict) -> tuple[str, bool]:
-        """Process a finding. Returns (message, is_duplicate)."""
-        # Resolve principle name for dedup key (prefer explicit, fall back to lookup)
-        p = args.get("p")
-        req = args.get("req")
-        if not p and req and req in self._reqs:
-            p = self._reqs[req]["principle"]
-        key = (p, args.get("file"), args.get("line"), args.get("t"))
-        if key in self._seen:
-            return "Duplicate finding, already captured. Move on.", True
-        self._seen.add(key)
-
-        finding: dict = {"schema_version": _FINDING_SCHEMA_VERSION}
-        finding.update({k: v for k, v in args.items() if v is not None})
-        self._enrich(args, finding)
-        self._enrich_code(finding)
-
-        self._fh.write(json.dumps(finding) + "\n")
-        self._fh.flush()
-        self.counter += 1
-        return f"Finding #{self.counter} recorded.", False
-
-
-_STOP_WORDS = frozenset({"a", "an", "the", "of", "to", "in", "for", "is", "and", "or", "not", "with", "without"})
-
-
-def _text_overlap(ref_name: str, description: str, reason: str) -> int:
-    """Score how well a ref name matches the finding text by counting shared words."""
-    stop = _STOP_WORDS
-    ref_words = set(ref_name.lower().split()) - stop
-    finding_words = (set(description.lower().split()) | set(reason.lower().split())) - stop
-    return len(ref_words & finding_words)
-
-
-def _select_best_refs(
-    all_refs: list[dict], description: str, reason: str,
-) -> list[dict]:
-    """Pick one ref per source type (CWE, CISQ, etc.), choosing the best text match.
-
-    When no words overlap at all, picks the broadest (first/lowest-ID) ref as a
-    safe default rather than an arbitrary specific one.
-    """
-    by_source: dict[str, list[dict]] = {}
-    for ref in all_refs:
-        source = ref.get("source", "") or ref.get("label", "").split("-")[0]
-        by_source.setdefault(source, []).append(ref)
-
-    result: list[dict] = []
-    for source, refs in by_source.items():
-        if len(refs) == 1:
-            result.append(refs[0])
-        else:
-            scored = [(r, _text_overlap(r.get("name", ""), description, reason)) for r in refs]
-            max_score = max(s for _, s in scored)
-            if max_score == 0:
-                # No text match -- pick the broadest (first listed, typically the parent)
-                result.append(refs[0])
-            else:
-                result.append(max(scored, key=lambda x: x[1])[0])
-    return result
+# Re-export public API so existing imports keep working.
+from quodeq.analysis.mcp.router import (  # noqa: F401
+    CompiledContext,
+    DeduplicationStore,
+    FileReader,
+    FindingsRouter,
+)
 
 
 def _build_compiled_context(sa: ServerArgs) -> CompiledContext:
